Chunk size: change from 1-MiB based to 1-M element based
This change of chunk size setting is to let allow datasets in a group to
have the same chunk dimension size along the first dimension (most
significant one). If evt.seq is used to calculated the partitioning
boundaries for all processes, the all datasets in the group can use the
same partitioning boundaries. Note datasets in a group can be of
different data types, which makes the 1-MiB based chunk size a poor
choice for parallel reads (hard to implement a consistent chunk
partitioning).