Clarify the limitation of input files
Input files are no longer required to have the same run IDs.
The subrun IDs in an input file are not required to be the same.
The partition key datasets are constructed and created based only on the
dataset provided in command-line option -k. The values of run and subrun
IDs are not used to calculate the key datasets. Concatenation is file
based, so the values of partition base datasets in one file are treated
larger than the files before it. Users are responsible to arrange the
input file names in order to achieve a desired order of datasets in the
concatenated output file.